fixes location of mpi tests
